splitting util.py
