+++ trulens_eval/trulens_eval/db.py
@@ -27,9 +27,9 @@ from trulens_eval.schema import FeedbackResultStatus
 from trulens_eval.schema import Perf
 from trulens_eval.schema import Record
 from trulens_eval.schema import RecordID
-from trulens_eval.util import JSON
-from trulens_eval.util import json_str_of_obj
-from trulens_eval.util import SerialModel
+from trulens_eval.utils.serial import JSON
+from trulens_eval.utils.serial import json_str_of_obj
+from trulens_eval.utils.serial import SerialModel
 from trulens_eval.utils.text import UNICODE_CHECK
 from trulens_eval.utils.text import UNICODE_CLOCK
