+++ trulens_eval/trulens_eval/feedback/feedback.py
@@ -22,13 +22,13 @@ from trulens_eval.schema import FeedbackResultID
 from trulens_eval.schema import FeedbackResultStatus
 from trulens_eval.schema import Record
 from trulens_eval.schema import Select
-from trulens_eval.util import FunctionOrMethod
-from trulens_eval.util import JSON
-from trulens_eval.util import jsonify
-from trulens_eval.util import TP
+from trulens_eval.utils.pyschema import FunctionOrMethod
+from trulens_eval.utils.serial import JSON
+from trulens_eval.utils.serial import jsonify
 from trulens_eval.utils.text import UNICODE_CHECK
 from trulens_eval.utils.text import UNICODE_CLOCK
 from trulens_eval.utils.text import UNICODE_YIELD
+from trulens_eval.utils.threading import TP
 
 logger = logging.getLogger(__name__)
