+++ trulens_eval/trulens_eval/instruments.py
@@ -274,13 +274,13 @@ from trulens_eval.schema import Perf
 from trulens_eval.schema import Query
 from trulens_eval.schema import RecordAppCall
 from trulens_eval.schema import RecordAppCallMethod
-from trulens_eval.util import _safe_getattr
-from trulens_eval.util import dict_merge_with
-from trulens_eval.util import get_all_local_in_call_stack
-from trulens_eval.util import get_first_local_in_call_stack
-from trulens_eval.util import jsonify
-from trulens_eval.util import JSONPath
-from trulens_eval.util import Method
+from trulens_eval.utils.containers import dict_merge_with
+from trulens_eval.utils.pyschema import Method
+from trulens_eval.utils.python import get_all_local_in_call_stack
+from trulens_eval.utils.python import get_first_local_in_call_stack
+from trulens_eval.utils.serial import _safe_getattr
+from trulens_eval.utils.serial import jsonify
+from trulens_eval.utils.serial import JSONPath
 
 logger = logging.getLogger(__name__)
 pp = PrettyPrinter()
