+++ trulens_eval/trulens_eval/tru_basic_app.py
@@ -2,22 +2,18 @@
 # Basic input output instrumentation and monitoring.
 """
 
-from datetime import datetime
-from inspect import BoundArguments, Signature
+from inspect import BoundArguments
+from inspect import Signature
 import logging
 from pprint import PrettyPrinter
-from typing import Callable, ClassVar, Sequence
+from typing import Callable, ClassVar
 
 from pydantic import Field
 
 from trulens_eval.app import App
 from trulens_eval.instruments import Instrument
-from trulens_eval.feedback.provider.endpoint import Endpoint
-from trulens_eval.schema import Cost
-from trulens_eval.schema import RecordAppCall
-from trulens_eval.util import Class
-from trulens_eval.util import FunctionOrMethod
-
+from trulens_eval.utils.pyschema import Class
+from trulens_eval.utils.pyschema import FunctionOrMethod
 
 logger = logging.getLogger(__name__)
 
@@ -60,7 +56,8 @@ class TruBasicApp(App):
     app: TruWrapperApp
 
     root_callable: ClassVar[FunctionOrMethod] = Field(
-        default_factory=lambda: FunctionOrMethod.of_callable(TruWrapperApp._call),
+        default_factory=lambda: FunctionOrMethod.
+        of_callable(TruWrapperApp._call),
         const=True
     )
 
@@ -91,7 +88,7 @@ class TruBasicApp(App):
     ) -> str:
         if "input" in bindings.arguments:
             return bindings.arguments['input']
-        
+
         return super().main_input(func, sig, bindings)
 
     def call_with_record(self, input: str, **kwargs):
